requirements.txt
requests>=2.28
opencv-python>=4.8
pyserial>=3.5
